+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Remove redundant code in `sd`.
 - Remove redundant code in `saw`.
+- Add Julia v1.12 to GitHub workflows.
 
 
 ### 0.8.2